Move `raw_window_handle` stuff to it's own mod / file
