+++ crates/wolf_engine_window/src/raw_window_handle.rs
@@ -1,0 +1,28 @@
+#[cfg(feature = "rwh_05")]
+pub use rwh_05;
+
+#[cfg(feature = "rwh_06")]
+pub use rwh_06;
+
+#[cfg(feature = "rwh_06")]
+pub trait HasRwh6Handles: rwh_06::HasWindowHandle + rwh_06::HasDisplayHandle {}
+#[cfg(feature = "rwh_06")]
+impl<T> HasRwh6Handles for T where T: rwh_06::HasWindowHandle + rwh_06::HasDisplayHandle {}
+
+#[cfg(not(feature = "rwh_06"))]
+pub trait HasRwh6Handles {}
+#[cfg(not(feature = "rwh_06"))]
+impl<T> HasRwh6Handles for T {}
+
+#[cfg(feature = "rwh_05")]
+pub trait HasRwh5Handles: rwh_05::HasRawWindowHandle + rwh_05::HasRawDisplayHandle {}
+#[cfg(feature = "rwh_05")]
+impl<T> HasRwh5Handles for T where T: rwh_05::HasRawWindowHandle + rwh_05::HasRawDisplayHandle {}
+
+#[cfg(not(feature = "rwh_05"))]
+pub trait HasRwh5Handles {}
+#[cfg(not(feature = "rwh_05"))]
+impl<T> HasRwh5Handles for T {}
+
+pub trait HasRawWindowHandles: HasRwh6Handles + HasRwh5Handles {}
+impl<T> HasRawWindowHandles for T where T: HasRwh6Handles + HasRwh5Handles {}
